Applying Modular Architecture
Applying modular architecture for better maintenance and feature creation
