+++ src/main/java/com/abacatepay/Models/Billing/BillingKind.java
@@ -1,0 +1,5 @@
+package com.abacatepay.Models.Billing;
+
+public enum BillingKind {
+    ONE_TIME
+}
