+++ src/main/java/com/abacatepay/Models/Billing/BillingMethod.java
@@ -1,0 +1,5 @@
+package com.abacatepay.Models.Billing;
+
+public enum BillingMethod {
+    PIX
+}
