+++ src/main/java/com/abacatepay/Models/Billing/BillingStatus.java
@@ -1,0 +1,9 @@
+package com.abacatepay.Models.Billing;
+
+public enum BillingStatus {
+    PENDING,
+    EXPIRED,
+    CANCELLED,
+    PAID,
+    REFUNDED
+}
