+++ src/main/java/com/abacatepay/Models/Billing/CreateBillingResponse.java
@@ -1,0 +1,20 @@
+package com.abacatepay.Models.Billing;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class CreateBillingResponse {
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String error;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private Billing billing;
+
+    public CreateBillingResponse(String error) {
+        this.error = error;
+    }
+}
