+++ src/main/java/com/abacatepay/Models/Billing/CustomerMetadata.java
@@ -1,0 +1,11 @@
+package com.abacatepay.Models.Billing;
+
+import lombok.Data;
+
+@Data
+public class CustomerMetadata {
+    private String name;
+    private String cellphone;
+    private String email;
+    private String taxId;
+}
