+++ src/main/java/com/abacatepay/Models/Client/exceptions/handlers/ApiExceptionModel.java
@@ -1,0 +1,8 @@
+package com.abacatepay.Models.Client.exceptions.handlers;
+
+import lombok.Data;
+
+@Data
+public class ApiExceptionModel {
+    private String error;
+}
