+++ src/main/java/com/abacatepay/Models/IAbacatePay.java
@@ -1,0 +1,5 @@
+package com.abacatepay.Models;
+
+public interface IAbacatePay {
+    IAbacatePayBilling billing();
+}
