+++ src/main/java/com/abacatepay/Models/IAbacatePayBilling.java
@@ -1,0 +1,10 @@
+package com.abacatepay.Models;
+
+import com.abacatepay.Models.Billing.CreateBillingData;
+import com.abacatepay.Models.Billing.CreateBillingResponse;
+import com.abacatepay.Models.Billing.ListBillingResponse;
+
+public interface IAbacatePayBilling {
+    CreateBillingResponse create(CreateBillingData data);
+    ListBillingResponse list();
+}
