--- src/main/java/com/abacatepay/clients/exceptions/FeignCustomException.java
@@ -1,10 +1,0 @@
-package com.abacatepay.clients.exceptions;
-
-import feign.FeignException;
-import feign.Request;
-
-public class FeignCustomException extends FeignException {
-    public FeignCustomException(Integer status,String message,Request request) {
-        super(status, message, request);
-    }
-}
