--- src/main/java/com/abacatepay/clients/exceptions/handlers/ApiExceptionModel.java
@@ -1,8 +1,0 @@
-package com.abacatepay.clients.exceptions.handlers;
-
-import lombok.Data;
-
-@Data
-public class ApiExceptionModel {
-    private String error;
-}
