--- src/main/java/com/abacatepay/model/IAbacatePay.java
@@ -1,5 +1,0 @@
-package com.abacatepay.model;
-
-public interface IAbacatePay {
-    IAbacatePayBilling billing();
-}
