--- src/main/java/com/abacatepay/model/IAbacatePayBilling.java
@@ -1,10 +1,0 @@
-package com.abacatepay.model;
-
-import com.abacatepay.model.billing.CreateBillingData;
-import com.abacatepay.model.billing.CreateBillingResponse;
-import com.abacatepay.model.billing.ListBillingResponse;
-
-public interface IAbacatePayBilling {
-    CreateBillingResponse create(CreateBillingData data);
-    ListBillingResponse list();
-}
