--- src/main/java/com/abacatepay/model/billing/BillingKind.java
@@ -1,5 +1,0 @@
-package com.abacatepay.model.billing;
-
-public enum BillingKind {
-    ONE_TIME
-}
