--- src/main/java/com/abacatepay/model/billing/BillingMethod.java
@@ -1,5 +1,0 @@
-package com.abacatepay.model.billing;
-
-public enum BillingMethod {
-    PIX
-}
