--- src/main/java/com/abacatepay/model/billing/BillingStatus.java
@@ -1,9 +1,0 @@
-package com.abacatepay.model.billing;
-
-public enum BillingStatus {
-    PENDING,
-    EXPIRED,
-    CANCELLED,
-    PAID,
-    REFUNDED
-}
