--- src/main/java/com/abacatepay/model/billing/CreateBillingData.java
@@ -1,16 +1,0 @@
-package com.abacatepay.model.billing;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Builder
-public class CreateBillingData {
-    private BillingKind frequency;
-    private List<BillingMethod> methods;
-    private List<CreateBillingProduct> products;
-    private String returnUrl;
-    private String completionUrl;
-}
