--- src/main/java/com/abacatepay/model/billing/CreateBillingResponse.java
@@ -1,20 +1,0 @@
-package com.abacatepay.model.billing;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class CreateBillingResponse {
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private String error;
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private Billing billing;
-
-    public CreateBillingResponse(String error) {
-        this.error = error;
-    }
-}
