--- src/main/java/com/abacatepay/model/billing/CustomerMetadata.java
@@ -1,11 +1,0 @@
-package com.abacatepay.model.billing;
-
-import lombok.Data;
-
-@Data
-public class CustomerMetadata {
-    private String name;
-    private String cellphone;
-    private String email;
-    private String taxId;
-}
