--- src/main/java/com/abacatepay/model/billing/ListBillingResponse.java
@@ -1,21 +1,0 @@
-package com.abacatepay.model.billing;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-public class ListBillingResponse {
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private String error;
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private List<Billing> billings;
-
-    public ListBillingResponse(String error) {
-        this.error = error;
-    }
-}
