--- src/test/java/com/abacatepay/clients/exceptions/handlers/ApiExceptionModelTest.java
@@ -1,30 +1,0 @@
-package com.abacatepay.clients.exceptions.handlers;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class ApiExceptionModelTest {
-
-    @Test
-    void testErrorField() {
-        ApiExceptionModel apiException = new ApiExceptionModel();
-        String expectedError = "Some error occurred";
-
-        apiException.setError(expectedError);
-        String actualError = apiException.getError();
-
-        assertEquals(expectedError, actualError, "Error field must be set correctly");
-    }
-
-    @Test
-    void testToString() {
-        ApiExceptionModel apiException = new ApiExceptionModel();
-        String expectedError = "Some error occurred";
-        apiException.setError(expectedError);
-
-        String actualString = apiException.toString();
-
-        assertEquals("ApiExceptionModel(error=Some error occurred)", actualString, "toString must return the correct string representation");
-    }
-}
